+++ src/client/components/tm_entity_properties_pane/tm_entity_properties_pane.js
@@ -60,51 +60,5 @@ Template.tm_entity_properties_pane.helpers({
     // for label
     itFor( label ){
         return 'entity_properties_'+label;
-    },
-
-    label(){
-        return this.item.get().label || '';
     }
-});
-
-Template.tm_entity_properties_pane.events({
-    // edit the managers
-    'click .js-edit-managers'( event, instance ){
-        const self = this;
-        Modal.run({
-            mdBody: 'accounts_select',
-            mdButtons: [{ id: Modal.C.Button.NEW, classes: 'btn-outline-primary me-auto' }, Modal.C.Button.CANCEL, { id: Modal.C.Button.OK, type: 'submit' }],
-            mdClasses: 'modal-lg',
-            mdClassesContent: Meteor.TM.Pages.current.page().get( 'theme' ),
-            mdTitle: pwixI18n.label( I18N, 'organizations.properties.managers_modal' ),
-            selected: self.item.get().DYN.managers || [],
-            update( selected ){
-                // get the item updated
-                self.item.get().DYN.managers = selected;
-                // get the form updated
-                instance.TM.form.get().setField( 'managers', self.item.get());
-                // advertise parents
-                instance.$( '.c-organization-properties-pane' ).trigger( 'panel-data', {
-                    emitter: 'managers',
-                    id: self.vtpid,
-                    ok: true,
-                    data: selected
-                });
-            }
-        });
-    },
-/*
-    // input checks
-    'input .c-organization-properties-pane'( event, instance ){
-        const dataContext = this;
-        if( !Object.keys( event.originalEvent ).includes( 'FormChecker' ) || event.originalEvent['FormChecker'].handled !== true ){
-            instance.TM.form.get().inputHandler( event ).then(( valid ) => { instance.TM.sendPanelData( dataContext, valid ); });
-        }
-    },
-
-    // ask for clear the panel
-    'iz-clear-panel .c-organization-properties-pane'( event, instance ){
-        instance.TM.form.get().clear();
-    }
-        */
 });